+++ stls.h
@@ -89,7 +89,7 @@ void write_guess_stls(double *SS, double *GG, input in);
 void read_guess_stls(double *SS, double *GG, input in);
 
 void check_guess_stls(int nx, double dx, double xmax, input in,
-		      int it_read, int it_expected, FILE *fid,
+		      size_t it_read, size_t it_expected, FILE *fid,
 		      bool check_grid, bool check_items, bool check_eof);
 
 #endif
